% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WavmatND.R
\name{dilate_filter}
\alias{dilate_filter}
\title{Dilate Filter}
\usage{
dilate_filter(filt, k)
}
\arguments{
\item{filt}{A vector representing the filter to be dilated.}

\item{k}{An integer specifying the step size for dilation.}
}
\value{
A vector of the dilated filter.
}
\description{
Inserts zeros between the elements of the input filter to dilate it.
}
\examples{
filt <- c(0.5, -0.5)
dilated_filt <- dilate_filter(filt, 2)
}
