% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WavmatND.R
\name{imbed}
\alias{imbed}
\title{Imbed Vector}
\usage{
imbed(a)
}
\arguments{
\item{a}{A vector to be embedded.}
}
\value{
A vector with zeros appended to double its length.
}
\description{
Embeds a vector by appending zeros to double its length.
}
\examples{
a <- c(1, 2, 3)
imbedded_a <- imbed(a)
}
