% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WavmatND.R
\name{repeating}
\alias{repeating}
\title{Repeat Vector}
\usage{
repeating(a, n)
}
\arguments{
\item{a}{A vector to be repeated.}

\item{n}{An integer specifying the number of times to repeat the vector.}
}
\value{
A vector with repeated elements.
}
\description{
Repeats the input vector a specified number of times.
}
\examples{
a <- c(1, 2)
repeated_a <- repeating(a, 3)
}
