% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WavmatWP.R
\name{WavPackMatWP}
\alias{WavPackMatWP}
\title{Construct Wavelet Packet Transformation Matrix}
\usage{
WavPackMatWP(h, N, k0, shift = 2)
}
\arguments{
\item{h}{Numeric vector representing the low-pass filter.}

\item{N}{Integer specifying the size of the matrix. Must be a power of 2.}

\item{k0}{Integer specifying the depth of the wavelet transformation. Should be between 1 and log2(N).}

\item{shift}{Integer for shifts in the wavelet transformation (default is 2).}
}
\value{
A matrix representing the wavelet packet transformation.
}
\description{
Construct Wavelet Packet Transformation Matrix
}
\examples{
# Example: Constructing a Wavelet Packet Transformation Matrix
h <- c(0.5, 0.5)  # Define a low-pass filter
N <- 8 # Size of the matrix (must be a power of 2)
k0 <- 3 # Depth of the wavelet transformation
shift <- 2 # Optional shift parameter
WP <- WavPackMatWP(h, N, k0, shift) # Construct the wavelet packet matrix
print(WP) # Display the resulting wavelet packet transformation matrix
}
