% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WavmatND.R
\name{WavmatND}
\alias{WavmatND}
\title{Wavelet Matrix Construction}
\usage{
WavmatND(hf, N, k, shift)
}
\arguments{
\item{hf}{A vector representing the high-pass filter.}

\item{N}{An integer specifying the size of the wavelet matrix.}

\item{k}{An integer indicating the number of iterations.}

\item{shift}{An integer representing the shift to be applied.}
}
\value{
A matrix representing the wavelet transformation.
}
\description{
Constructs a wavelet matrix using a high-pass filter, specified parameters, and an iterative process to build matrices.
}
\examples{
# Example: Constructing a Wavelet Matrix
hf <- c(0.5, -0.5) # Define a high-pass filter
N <- 8             # Size of the wavelet matrix
k <- 3             # Number of iterations
shift <- 2         # Apply a shift
W <- WavmatND(hf, N, k, shift) # Construct the wavelet matrix
print(W)           # Display the resulting wavelet matrix
}
