% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wavmat.R
\name{WavPackMat}
\alias{WavPackMat}
\title{Construct Highly Optimized Orthogonal Wavelet Transformation Matrix}
\usage{
WavPackMat(h, N, k0, shift = 2)
}
\arguments{
\item{h}{Numeric vector representing the low-pass filter.}

\item{N}{Integer specifying the size of the matrix. Must be a power of 2.}

\item{k0}{Integer specifying the depth of the wavelet transformation. Should be between 1 and log2(N).}

\item{shift}{Integer for shifts in the wavelet transformation (default is 2).}
}
\value{
A sparse matrix representing the wavelet packet transformation.
}
\description{
This function creates an optimized wavelet transformation matrix for high-quality filter coefficients
and sparse matrix operations.
}
