% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WavmatWP.R
\name{WavPackMatWP}
\alias{WavPackMatWP}
\title{Construct Wavelet Packet Transformation Matrix}
\usage{
WavPackMatWP(h, N, k0, shift = 2)
}
\arguments{
\item{h}{Numeric vector representing the low-pass filter.}

\item{N}{Integer specifying the size of the matrix. Must be a power of 2.}

\item{k0}{Integer specifying the depth of the wavelet transformation. Should be between 1 and log2(N).}

\item{shift}{Integer for shifts in the wavelet transformation (default is 2).}
}
\value{
A matrix representing the wavelet packet transformation.
}
\description{
Construct Wavelet Packet Transformation Matrix
}
